# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library Management API",
    "version": "1.0"
  },
  "paths": {
    "/books": {
      "post": {
        "summary": "Add a new book",
        "parameters": [
          {
            "name": "book",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Title": { "type": "string" },
                "Author": { "type": "string" },
                "Published Year": { "type": "integer" },
                "ISBN": { "type": "string" },
                "Genre": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Book added" }
        }
      },
      "get": {
        "summary": "List all books",
        "responses": {
          "200": { "description": "A list of books" }
        }
      }
    },
    "/books/search": {
      "get": {
        "summary": "Search for books",
        "parameters": [
          {
            "name": "author",
            "in": "query",
            "type": "string"
          },
          {
            "name": "year",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "genre",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "A list of filtered books" }
        }
      }
    },
    "/books/{isbn}": {
      "delete": {
        "summary": "Delete a book by ISBN",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": { "description": "Book deleted" }
        }
      },
      "put": {
        "summary": "Update book details",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "book",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Title": { "type": "string" },
                "Author": { "type": "string" },
                "Published Year": { "type": "integer" },
                "ISBN": { "type": "string" },
                "Genre": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Book updated" },
          "404": { "description": "Book not found" }
        }
      }
    }
  }
}